what eve

+++ inc/Collision.h
@@ -8,14 +8,14 @@ using namespace Eigen;
 
 
 
-tuple<bool, Vector3f> CD_bowl(MatrixXf vertices);
+tuple<bool, Vector3d> CD_bowl(MatrixXd vertices);
 
-tuple<bool, Vector3f> CD_table(MatrixXf vertices);
+tuple<bool, Vector3d> CD_table(MatrixXd vertices);
 
-tuple<bool, Vector3f> CD_bowl_wide(MatrixXf vertices);
+tuple<bool, Vector3d> CD_bowl_wide(MatrixXd vertices);
 
-tuple<bool, Vector3f> CD_vshape(MatrixXf vertices);
+tuple<bool, Vector3d> CD_vshape(MatrixXd vertices);
 
-tuple<bool, vector<int>> CD_table_FEM(MatrixXf vertices);
+tuple<bool, vector<int>> CD_table_FEM(MatrixXd vertices);
 
 #endif
